+++ packages/client/src/pages/game/game.tsx
@@ -1,4 +1,5 @@
 import React, { FC, useRef, useEffect } from 'react';
+import renderHorizontalText from './hame.helper';
 import style from './game.module.scss';
 import style from './canvas.module.scss';
 import React, { FC, MouseEventHandler } from 'react';
@@ -73,57 +74,11 @@ const Game: FC<TGame> = props => {
 
             // render horisontal text
             // eslint-disable-next-line
-            function fillTextWithSpacing(context: CanvasRenderingContext2D, text: string, x: number, y: number, spacing: number) {
-                const totalWidth = context.measureText(text).width + spacing * (text.length - 1);
-
-                const align = context.textAlign;
-                context.textAlign = 'left';
-                /* eslint-disable */
-                switch (align) {
-                    case 'right':
-                        x -= totalWidth;
-                        break;
-                    case 'center':
-                        x -= totalWidth / 2;
-                        break;
-                }
-                /* eslint-enable */
-
-                let offset;
-                let pairWidth;
-                let charWidth;
-                let charNextWidth;
-                let pairSpacing;
-                let char;
-                let charNext;
-
-                // eslint-disable-next-line
-                for (offset = 0; offset < text.length; offset = offset + 1) {
-                    char = text.charAt(offset);
-                    pairSpacing = 0;
-                    if (offset + 1 < text.length) {
-                        charNext = text.charAt(offset + 1);
-
-                        pairWidth = context.measureText(char + charNext).width;
-                        charWidth = context.measureText(char).width;
-                        charNextWidth = context.measureText(charNext).width;
-                        pairSpacing = pairWidth - charWidth - charNextWidth;
-                    }
-
-                    context.fillText(char, x, y);
-                    if (charWidth) {
-                        // eslint-disable-next-line
-                        x = x + pairSpacing + spacing + charWidth;
-                    }
-                }
-
-                context.textAlign = align;
-            }
             ctx.font = '19px Tektur';
             ctx.fillStyle = 'white';
-            const textH = 'A B C D E F G H I G';
+            const textHeight = 'A B C D E F G H I G';
             // eslint-disable-next-line
-            [...Array(2).keys()].forEach(i => fillTextWithSpacing(ctx, textH, 257 + 400 * i, 63, 6.6));
+            [...Array(2).keys()].forEach(i => renderHorizontalText(ctx, textHeight, 257 + 400 * i, 63, 6.6));
 
             // render horizontal text
             // eslint-disable-next-line
